move function results hashing to lang
We need to abstract the function results verification to use internally
too, so start by moving it out of the providers code.
